Avoid overlong flag fields in TEvent
There is no possible 'eventFlags' or 'controlKeyState' flag that won't fit in 16 bits, even when copied straight from Win32's MOUSE_EVENT_RECORD or KEY_EVENT_RECORD. Replacing ulong with ushort reduces TEvent's size from 40 to 24 bytes in 64-bit Linux.

+++ include/tvision/system.h
@@ -77,8 +77,8 @@ const int meTripleClick = 0x10;
 struct MouseEventType
 {
     TPoint where;
-    ulong eventFlags;           // Replacement for doubleClick.
-    ulong controlKeyState;
+    ushort eventFlags;           // Replacement for doubleClick.
+    ushort controlKeyState;
     uchar buttons;
     uchar wheel;
 };
@@ -197,7 +197,7 @@ struct KeyDownEvent
         ushort keyCode;
         CharScanType charScan;
         };
-    ulong controlKeyState;
+    ushort controlKeyState;
     char text[4];
     uchar textLength;
